requirements.txt
starlette
uvicorn
graphene
mongoengine
graphene_mongo
websockets
starlette-jwt
pyjwt
aiohttp